local host webpage
This wraps up my project. This includes the "Processed_Bengaluru_Dataset.csv" which is a spreadsheet csv file (basically the dataset), the Python file where the code is written for the whole project including the webpage part, I attached a short video that shows how the webpage is working and "index.html", "column.json", "bengaluru.pickle" files. The first one is the webpage html code file and the later two are also used for the webpage creation. This concludes my project for "Real-Time House Price Prediction using Supervised Machine Learning".